Prune references in the mirror when fetching.
Consider a ref is deleted in the remote and then a new ref is created as a child (e.g. refs/heads/feature is deletede and refs/heads/feature/iteration is created).
If we don't prune we run into a problem: we cannot create the iteration ref file in the feature directory because feature is a ref file itself:

    error: cannot lock ref 'refs/heads/feature/iteration': 'refs/heads/feature' exists; cannot create 'refs/heads/feature/iteration'

So here we first delete all refs in the mirror which don't exist in the remote.